+++ tough/src/error.rs
@@ -358,6 +358,55 @@ pub enum Error {
         backtrace: Backtrace,
     },
 
+    #[cfg(feature = "http")]
+    #[snafu(display("Error creating HTTP client for '{}': {}", url, source))]
+    HttpClientBuild {
+        url: Url,
+        source: reqwest::Error,
+        backtrace: Backtrace,
+    },
+
+    #[cfg(feature = "http")]
+    #[snafu(display("Error fetching target from '{}': {}", url, source))]
+    HttpFetch {
+        url: Url,
+        source: reqwest::Error,
+        backtrace: Backtrace,
+    },
+
+    #[cfg(feature = "http")]
+    #[snafu(display(
+        "HTTP header '{}' contains invalid characters: {}",
+        header_value,
+        source
+    ))]
+    HttpHeader {
+        header_value: String,
+        source: reqwest::header::InvalidHeaderValue,
+        backtrace: Backtrace,
+    },
+
+    #[cfg(feature = "http")]
+    #[snafu(display("Error creating HTTP request for '{}': {}", url, source))]
+    HttpRequestBuild {
+        url: Url,
+        source: reqwest::Error,
+        backtrace: Backtrace,
+    },
+
+    #[cfg(feature = "http")]
+    #[snafu(display(
+        "Failed to fetch '{}' after {} tries, final error: {}",
+        url,
+        tries,
+        source
+    ))]
+    HttpRetries {
+        url: Url,
+        tries: u32,
+        source: reqwest::Error,
+    },
+
     #[snafu(display("Failed to walk directory tree '{}': {}", directory.display(), source))]
     WalkDir {
         directory: PathBuf,