+++ tough/CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Breaking Changes
+- The `HttpTransport` type and the `Read` and `Error` types that it uses have changed.
+
+### Added
+- Retries in HttpTransport.
+
+### Added
+- Added HTTP retry logic. #143
+
 ## [0.7.0] - 2020-06-26
 
 ### Added
@@ -57,6 +67,8 @@ For changes that require modification of calling code see #120 and #121.
 ### Added
 - Everything!
 
+[Unreleased]: https://github.com/awslabs/tough/compare/tough-v0.6.0...HEAD
+[0.6.0]: https://github.com/awslabs/tough/compare/tough-v0.5.0...tough-v0.6.0
 [0.5.0]: https://github.com/awslabs/tough/compare/tough-v0.4.0...tough-v0.5.0
 [0.4.0]: https://github.com/awslabs/tough/compare/tough-v0.3.0...tough-v0.4.0
 [0.3.0]: https://github.com/awslabs/tough/compare/tough-v0.2.0...tough-v0.3.0